requirements.txt
numpy==1.19.5
tensorflow==2.4.1
Keras==2.4.3
h5py==2.10.0
Pillow==8.2.0
opencv-python==4.5.1.48
scikit-image==0.17.2
